refactor legend state to mapsInformation state
