+++ features/geojson/Sidebar/Maps/Legend/ColorSwatch.tsx
