+++ features/geojson/Sidebar/Maps/Legend/Settings.tsx
