+++ features/geojson/Sidebar/Maps/index.tsx
@@ -1,0 +1,11 @@
+import { Legend } from "./Legend";
+import { Divider } from "@components/Divider";
+
+export function Maps() {
+  return (
+    <>
+      <Divider label="Legend and Color Generator" />
+      <Legend />
+    </>
+  );
+}
