+++ features/geojson/store/types.ts
@@ -39,7 +39,7 @@ export interface GeoJSONExtended extends FeatureCollection {
 
 
 /**
- * GeoJSON state and function for handle features collection data
+ * Data state and function for handle features collection data
  * 
  */
 export interface DataState extends Omit<GeoJSONExtended, "type"> {
@@ -72,8 +72,8 @@ export interface DataFunction {
 export type DataStore = DataState & DataFunction
 
 /**
- * Legend state and function for handle legend data and 
- * color generator. 
+ * Maps Information state and function for handle legend and color,
+ * proportional circle, and and label
  * 
  */
 export type LegendOptions = {
@@ -142,7 +142,7 @@ export interface ProportionalCircle {
   borderColor: string
 }
 
-export interface LegendState {
+export interface MapInformationState {
   legends: Legend[];
 
   legendTitle: string;
@@ -154,7 +154,7 @@ export interface LegendState {
   proportionalCircle: ProportionalCircle
 }
 
-export interface LegendFunction {
+export interface MapInformationFunction {
   addLegends: () => void;
 
   updateLegend: (uuid: string, legend: Legend) => void;
@@ -184,7 +184,7 @@ export interface LegendFunction {
   updateProportionalCircle: (settings: Partial<ProportionalCircle>) => void
 }
 
-export type LegendStore = LegendState & LegendFunction
+export type MapInformationStore = MapInformationState & MapInformationFunction
 
 
 /**
@@ -223,8 +223,6 @@ export interface SettingsState {
 
   baseMap: BaseMap | null;
 
-  showMapControls: boolean;
-
   geoJSONSettings: GeoJSONSettings
 
 }
@@ -242,8 +240,6 @@ export interface SettingsFunction {
 
   downloadMap: (format?: "png" | "jpeg" | "svg") => void
 
-  toggleMapControls: () => void;
-
   updateGeoJSONSettings: (settings: Partial<GeoJSONSettings>) => void;
 }
 
@@ -254,4 +250,4 @@ export type SettingsStore = SettingsState & SettingsFunction
  * All combined state
  * 
  */
-export type Store = DataStore & LegendStore & SettingsStore
+export type Store = DataStore & MapInformationStore & SettingsStore
